Issue 1096
