+++ app/src/components/proposals/proposalsPage/components/Proposals.tsx
@@ -1,12 +1,7 @@
-import ProposalCard, {
-  ProposalArrow,
-  ProposalBadge,
-  ProposalBadgeType,
-  ProposalDates,
-  ProposalTextContent,
-} from "@/components/proposals/proposalsPage/components/ProposalCard"
+import { ProposalBadgeType } from "@/components/proposals/proposalsPage/components/ProposalCard"
+import { ProposalRow } from "./ProposalRow"
 
-interface StandardProposalProps {
+export interface StandardProposalProps {
   rounded?: boolean
   voted?: boolean
   passed?: boolean
@@ -24,52 +19,6 @@ interface StandardProposalProps {
   arrow: {
     href: string
   }
-}
-const ProposalRow = (props: StandardProposalProps) => {
-  const voteStatus = () => {
-    if (props.badge.badgeType === ProposalBadgeType.now) {
-      if (props.voted) {
-        return {
-          text: "You voted",
-          styling: "text-success-foreground",
-        }
-      }
-      return {
-        text: "You haven't voted yet",
-        styling: "text-primary",
-      }
-    } else if (props.badge.badgeType === ProposalBadgeType.past) {
-      if (props.passed) {
-        return {
-          text: "Result Positive ie: Passed",
-          styling: "text-success-foreground",
-        }
-      }
-      return {
-        text: "Result Negative ie: Failed",
-        styling: "text-primary",
-      }
-    }
-    return undefined
-  }
-
-  return (
-    <ProposalCard rounded={props.rounded || false}>
-      <ProposalBadge type={props.badge.badgeType} />
-      <ProposalTextContent
-        title={props.textContent.title}
-        subtitle={props.textContent.subtitle}
-      />
-      <div className="hidden md:block">
-        <ProposalDates
-          startDate={props.dates.startDate}
-          endDate={props.dates.endDate}
-          voteStatus={voteStatus()}
-        />
-      </div>
-      <ProposalArrow href={props.arrow.href} />
-    </ProposalCard>
-  )
 }
 interface StandardProposalsProps {
   proposals: StandardProposalProps[]
